+++ client-side-js/allControls.js
@@ -1,4 +1,4 @@
-async function clientSide_getControl(controlSelector) {
+async function clientSide_allControls(controlSelector) {
     controlSelector = await Promise.resolve(controlSelector) // to plug into fluent async api
     return await browser.executeAsync((controlSelector, done) => {
         const errorHandling = (error) => {
@@ -10,22 +10,28 @@ async function clientSide_getControl(controlSelector) {
         if (controlSelector.timeout) {
             waitForUI5Options.timeout = controlSelector.timeout
         }
+
         window.wdi5.waitForUI5(
             waitForUI5Options,
             () => {
                 window.wdi5.Log.info("[browser wdi5] locating " + JSON.stringify(controlSelector))
                 controlSelector.selector = window.wdi5.createMatcher(controlSelector.selector)
                 window.bridge
-                    .findDOMElementByControlSelector(controlSelector)
-                    .then((domElement) => {
+                    .findAllDOMElementsByControlSelector(controlSelector)
+                    .then((domElements) => {
                         // window.wdi5.Log.info('[browser wdi5] control located! - Message: ' + JSON.stringify(domElement));
                         // ui5 control
-                        const ui5Control = window.wdi5.getUI5CtlForWebObj(domElement)
-                        const id = ui5Control.getId()
-                        window.wdi5.Log.info(`[browser wdi5] control with id: ${id} located!`)
-                        const aProtoFunctions = window.wdi5.retrieveControlMethods(ui5Control)
-                        // @type [String, String?, String, "Array of Strings"]
-                        done(["success", domElement, id, aProtoFunctions])
+                        let returnElements = []
+                        domElements.forEach((domElement) => {
+                            const ui5Control = window.wdi5.getUI5CtlForWebObj(domElement)
+                            const id = ui5Control.getId()
+                            window.wdi5.Log.info(`[browser wdi5] control with id: ${id} located!`)
+                            const aProtoFunctions = window.wdi5.retrieveControlMethods(ui5Control)
+                            // @type [String, String?, String, "Array of Strings"]
+                            returnElements.push({ domElement: domElement, id: id, aProtoFunctions: aProtoFunctions })
+                        })
+
+                        done(["success", returnElements])
                     })
                     .catch(errorHandling)
             },
@@ -35,5 +41,5 @@ async function clientSide_getControl(controlSelector) {
 }
 
 module.exports = {
-    clientSide_getControl
+    clientSide_allControls
 }