+++ examples/ui5-js-app/wdio-selenium-service.conf.js
@@ -8,8 +8,10 @@ const _config = {
     },
     specs: [join("webapp", "test", "e2e", "**/*.test.js")],
     exclude: [join("webapp", "test", "e2e", "ui5-late.test.js")],
+    logLevel: "error",
     bail: 0,
-    baseUrl: "http://localhost:8888"
+    baseUrl: "http://localhost:8888",
+    services: [["selenium-standalone", { drivers: { chrome: true, chromiumedge: "latest" } }]]
 }
 
 exports.config = merge(baseConfig, _config)