+++ src/lib/wdi5-control.ts
@@ -20,15 +20,42 @@ export class WDI5Control {
     _webElement: WebdriverIO.Element | string = null
     _webdriverRepresentation: WebdriverIO.Element = null
     _wdio_ui5_key: string = null
-    _generatedUI5Methods: [] | string = null
+    _generatedUI5Methods: Array<string>
     _initialisation = false
     _forceSelect = false
+    _wdioBridge = <WebdriverIO.Element>{}
+    _generatedWdioMethods: Array<string>
+    _domId: string
 
-    constructor() {
+    constructor(oOptions) {
+        const {
+            controlSelector,
+            wdio_ui5_key,
+            forceSelect,
+            generatedUI5Methods,
+            webdriverRepresentation,
+            webElement,
+            domId
+        } = oOptions
+
+        this._controlSelector = controlSelector
+        this._wdio_ui5_key = wdio_ui5_key
+        this._forceSelect = forceSelect
+        this._generatedUI5Methods = generatedUI5Methods
+        this._webElement = webElement
+        this._webdriverRepresentation = webdriverRepresentation
+        this._domId = domId
+
+        this.attachControlBridge(this._generatedUI5Methods as Array<string>)
+        this.attachWdioControlBridge(this._generatedWdioMethods as Array<string>)
+
+        // set the succesful init param
+        this._initialisation = true
+
         return this
     }
 
-    async init(controlSelector, forceSelect) {
+    async init(controlSelector = this._controlSelector, forceSelect = this._forceSelect) {
         this._controlSelector = controlSelector
         this._wdio_ui5_key = controlSelector.wdio_ui5_key
         this._forceSelect = forceSelect
@@ -44,7 +71,8 @@ export class WDI5Control {
 
             // dynamic function bridge
             this._generatedUI5Methods = controlResult[1]
-            await this.attachControlBridge(this._generatedUI5Methods as Array<string>)
+            this.attachControlBridge(this._generatedUI5Methods as Array<string>)
+            this.attachWdioControlBridge(this._generatedWdioMethods as Array<string>)
 
             // set the succesful init param
             this._initialisation = true
@@ -61,9 +89,13 @@ export class WDI5Control {
     }
 
     /**
-     * @return the webdriver Element
+     * @return {WebdriverIO.Element} the webdriver Element
      */
     async getWebElement() {
+        if (!this._webdriverRepresentation) {
+            // to enable transition from wdi5 to wdio api in allControls
+            await this.renewWebElement()
+        }
         //// TODO: check this "fix"
         //// why is the renew necessary here?
         //// it causes hiccup with the fluent async api as the transition from node-scope
@@ -77,6 +109,23 @@ export class WDI5Control {
         } else {
             return this._webdriverRepresentation
         }
+    }
+
+    /**
+     * add conveniance to the getWebElement Function
+     * @returns {WebdriverIO.Element} the webdriver Element
+     */
+    $() {
+        return this._wdioBridge // this.getWebElement()
+    }
+
+    /**
+     * @param id
+     * @returns
+     */
+    async renewWebElement(id: string = this._domId) {
+        this._webdriverRepresentation = await $(`//*[@id="${id}"]`)
+        return this._webdriverRepresentation
     }
 
     /**
@@ -134,7 +183,7 @@ export class WDI5Control {
      * this method is also used wdi5-internally to implement the extended forceSelect option
      */
     async renewWebElementReference() {
-        const newWebElement = (await this.getControl())[0]
+        const newWebElement = (await this.getControl({ selector: { id: this._domId } }))[0] // added to have a more stable retrieval experience
         this._webElement = newWebElement
         return newWebElement
     }
@@ -207,11 +256,24 @@ export class WDI5Control {
      *
      * @param sReplFunctionNames
      */
-    private async attachControlBridge(sReplFunctionNames: Array<string>) {
+    private attachControlBridge(sReplFunctionNames: Array<string>) {
         // check the validity of param
         if (sReplFunctionNames) {
             sReplFunctionNames.forEach(async (sMethodName) => {
-                this[sMethodName] = await this.executeControlMethod.bind(this, sMethodName, this._webElement)
+                this[sMethodName] = this.executeControlMethod.bind(this, sMethodName, this._webElement)
+            })
+        } else {
+            Logger.warn(`${this._wdio_ui5_key} has no sReplFunctionNames`)
+        }
+    }
+
+    private attachWdioControlBridge(sReplFunctionNames: Array<string>) {
+        // check the validity of param
+        if (sReplFunctionNames) {
+            sReplFunctionNames.forEach(async (sMethodName) => {
+                this._wdioBridge[sMethodName] = async (): Promise<any> => {
+                    return await (await this.getWebElement())[sMethodName]()
+                }
             })
         } else {
             Logger.warn(`${this._wdio_ui5_key} has no sReplFunctionNames`)
@@ -394,11 +456,58 @@ export class WDI5Control {
         if (result[2]) {
             // only if the result is valid
             this._webdriverRepresentation = await $(`//*[@id="${result[2]}"]`)
+            this._generatedWdioMethods = this._retrieveControlMethods(this._webdriverRepresentation)
+
+            this._domId = result[2]
         }
 
         this.writeResultLog(result, "getControl()")
 
         return [result[1], result[3]]
+    }
+
+    /**
+     *
+     * @param {WebDriver.Element} control
+     * @returns {Array<string>}
+     */
+    private _retrieveControlMethods(control) {
+        const _control = control
+        // create keys of all parent prototypes
+        const properties = new Set()
+        do {
+            Object.getOwnPropertyNames(control).map((item) => properties.add(item))
+        } while ((control = Object.getPrototypeOf(control)))
+
+        // @ts-ignore
+        const controlMethodsToProxy = [...properties.keys()].filter((item: string) => {
+            if (typeof _control[item] === "function") {
+                // function
+
+                // filter private methods
+                if (item.startsWith("_")) {
+                    return false
+                }
+
+                // filter not working methods
+                // and those with a specific api from wdi5/wdio-ui5-service
+                const aFilterFunctions = ["$", "constructor"]
+
+                if (aFilterFunctions.includes(item)) {
+                    return false
+                }
+
+                /* if (item.startsWith("is")) {
+                    // only check functions
+                    return true
+                } */
+
+                return true
+            }
+            return false
+        })
+
+        return controlMethodsToProxy as Array<string>
     }
 
     /**